Updating structure and md
- Updated the folder structure to help importing packages
- Added dummy libqhyccd.so to make sure it passed the commit
- updated and added md files
