+++ src/ecal/ecal_util/image_message.py
