+++ src/ecal/ecal_util/jpeg_compression.py
