+++ src/ecal/proto_files/camera_image.proto
