+++ src/ecal/proto_files/web_camera_image.proto
